Data and entity schape
Relocated factories and real data to a separate package

Integrate new package structure types / data into application

More explicit where lint actions direct to, update dependencies
