+++ package.json
@@ -17,7 +17,7 @@
     "url": "https://github.com/project-millipede/millipede-docs/issues"
   },
   "scripts": {
-    "lint": "eslint './docs/**/*' './src/**/*' './pages/**/*'",
+    "lint": "eslint './docs/src/**/*' './src/**/*' './pages/**/*'",
     "next:dev": "next",
     "next:build": "next build",
     "next:start": "next start",
@@ -53,7 +53,7 @@
     "draft-js": "^0.11.4",
     "factory.ts": "^0.5.1",
     "faker": "^4.1.0",
-    "framer-motion": "^1.7.0",
+    "framer-motion": "^1.7.2",
     "gray-matter": "^4.0.2",
     "group-array": "^1.0.0",
     "houx": "^1.1.2",
@@ -62,7 +62,7 @@
     "lodash": "^4.17.15",
     "mdast-util-toc": "^5.0.1",
     "mui-rte": "^1.11.0",
-    "next": "^9.2.1-canary.2",
+    "next": "^9.2.1-canary.4",
     "next-i18next-serverless": "^1.1.72",
     "normalizr": "^3.4.1",
     "notistack": "^0.9.7",
@@ -104,8 +104,8 @@
     "@types/react-dom": "^16.9.4",
     "@types/styled-components": "^4.4.2",
     "@types/webpack-merge": "^4.1.5",
-    "@typescript-eslint/eslint-plugin": "^2.16.0",
-    "@typescript-eslint/parser": "^2.16.0",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "@typescript-eslint/parser": "^2.17.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "core-js": "^3.6.4",
