+++ src/data/social/index.ts
@@ -1,0 +1,5 @@
+import * as factories from './factories';
+import * as schema from './schema';
+import * as mocks from './mocks';
+
+export { factories, schema, mocks };
