+++ src/demo/social/components/Timeline.svc.ts
@@ -2,16 +2,14 @@ import uuid from 'uuid';
 
 import { TimelineActions } from '../../../../docs/src/modules/redux/features/actionType';
 import { createPost } from '../../../../docs/src/modules/redux/features/timeline/actions';
-import {
-  contentFactory,
-  currentTimeStamps,
-  denormalizeWrapper,
-  Post,
-  timelineSchema,
-  UseCaseEntities,
-  mediaFactory
-} from '../../../typings/social';
 
+import { Post } from '../../../typings/social';
+import { UseCaseEntities } from '../../../typings/social/schema';
+
+import { factories, schema } from '../../../data/social';
+
+const { contentFactory, currentTimeStampFactory, mediaFactory } = factories;
+const { denormalizeWrapper, timelineSchema } = schema;
 export const handleCreatePost = (
   timelineId: number,
   text: string,
@@ -30,7 +28,7 @@ export const handleCreatePost = (
 
     author: { id: -1 },
 
-    content: contentFactory.combine(currentTimeStamps).build(),
+    content: contentFactory.combine(currentTimeStampFactory).build(),
     comments: [],
     votes: []
   };
